+++ update-doi-handler/build.gradle
@@ -1,6 +1,7 @@
 def zalandoVersion = '0.23.0'
 
 dependencies {
+    api group: 'com.github.BIBSYSDEV', name: 'nva-datamodel-java', version: '0.10.6'
     implementation group: 'com.github.BIBSYSDEV', name: 'nva-commons', version: 'v0.3.3'
     implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
     implementation group: 'com.amazonaws', name: 'aws-lambda-java-core', version: '1.2.0'
