+++ update-doi-handler/src/test/java/no/unit/nva/doi/updatedoi/model/datacite/request/DraftDoiRequestTest.java
@@ -1,12 +1,12 @@
-package no.unit.nva.doi.assigndoi.model.datacite.request;
+package no.unit.nva.doi.updatedoi.model.datacite.request;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 
-import no.unit.nva.doi.assigndoi.model.datacite.request.DraftDoiRequest.Attributes;
-import no.unit.nva.doi.assigndoi.model.datacite.request.DraftDoiRequest.AttributesBuilder;
-import no.unit.nva.doi.assigndoi.model.datacite.request.DraftDoiRequest.Data;
+import no.unit.nva.doi.updatedoi.model.datacite.request.DraftDoiRequest.Attributes;
+import no.unit.nva.doi.updatedoi.model.datacite.request.DraftDoiRequest.AttributesBuilder;
+import no.unit.nva.doi.updatedoi.model.datacite.request.DraftDoiRequest.Data;
 import org.junit.jupiter.api.Test;
 
 class DraftDoiRequestTest {
